feat: 为数字类型的Shared实现各种计算符号

+++ winia/src/core.rs
@@ -4,12 +4,12 @@ use lazy_static::lazy_static;
 use std::sync::Mutex;
 
 lazy_static! {
-    static ref PRE_ID: Mutex<usize> = Mutex::new(0);
-    static ref STR_TO_ID: Mutex<HashMap<String, usize>> = Mutex::new(HashMap::new());
+    static ref PRE_ID: Mutex<u32> = Mutex::new(0);
+    static ref STR_TO_ID: Mutex<HashMap<String, u32>> = Mutex::new(HashMap::new());
 }
 
 /// Generate a unique id
-pub fn next_id() -> usize {
+pub fn next_id() -> u32 {
     let mut id = PRE_ID.lock().unwrap();
     *id += 1;
     *id
@@ -17,7 +17,7 @@ pub fn next_id() -> usize {
 
 /// Bind a string to an id
 /// Use [get_id_by_name] to get the id by the string
-pub fn bind_str_to_id(s: &str, id: usize) {
+pub fn bind_str_to_id(s: &str, id: u32) {
     if s.is_empty() {
         panic!("The string used to bind to an id cannot be empty");
     }
@@ -30,13 +30,13 @@ pub fn bind_str_to_id(s: &str, id: usize) {
 
 /// Get the id by the string
 /// Use [bind_str_to_id] to bind the string to an id
-pub fn get_id_by_name(s: impl AsRef<str>) -> Option<usize> {
+pub fn get_id_by_name(s: impl AsRef<str>) -> Option<u32> {
     let s = s.as_ref();
     let str_to_id = STR_TO_ID.lock().unwrap();
     str_to_id.get(s).map(|id| *id)
 }
 
-pub fn get_name_by_id(id: usize) -> Option<String> {
+pub fn get_name_by_id(id: u32) -> Option<String> {
     let str_to_id = STR_TO_ID.lock().unwrap();
     for (key, value) in str_to_id.iter() {
         if *value == id {
@@ -52,7 +52,7 @@ pub fn unbind_str_to_id(s: impl AsRef<str>) {
     str_to_id.remove(s);
 }
 
-pub fn unbind_id(id: usize) {
+pub fn unbind_id(id: u32) {
     let mut str_to_id = STR_TO_ID.lock().unwrap();
     let mut keys_to_remove = Vec::new();
     for (key, value) in str_to_id.iter() {
